Fixed Bug 10694: The new document size seems to be out of kilter.
git-svn-id: svn://scribus.net/trunk/Scribus@17465 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/ui/applytemplatedialog.cpp
@@ -171,7 +171,7 @@ void ApplyMasterPageDialog::setup(ScribusDoc *view, QString Nam)
 	fromPageSpinBox->setMaximum(docPagesCount);
 	fromPageSpinBox->setValue(view->currentPage()->pageNr()+1);
 	toPageSpinBox->setMaximum(docPagesCount);
-	toPageSpinBox->setValue(docPagesCount);
+	toPageSpinBox->setValue(static_cast<int>(docPagesCount));
 }
